+++ plugins/com.onpositive.prefeditor/src/com/onpositive/prefeditor/model/FolderPreferenceProvider.java
@@ -8,19 +8,89 @@ import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.file.FileSystems;
+import java.nio.file.Path;
+import java.nio.file.StandardWatchEventKinds;
+import java.nio.file.WatchEvent;
+import java.nio.file.WatchKey;
+import java.nio.file.WatchService;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Properties;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.TimeUnit;
+
+import org.eclipse.ui.services.IDisposable;
 
 import com.onpositive.prefeditor.PrefEditorPlugin;
 
-public class FolderPreferenceProvider implements IPreferenceProvider {
+public class FolderPreferenceProvider implements IPreferenceProvider, IDisposable {
 	
 	protected static final String EXT = ".prefs";
 	
 	protected Map<String, Properties> propFiles = new HashMap<>();
 	
-	protected File preferenceFolder;
+	protected WatchService watcher; 
+	
+	protected ScheduledExecutorService scheduledExecutorService = Executors.newScheduledThreadPool(1);
+	
+	final protected File preferenceFolder;
+
+	private boolean tracking;
+
+	private Runnable watchRunnable = new Runnable() {
+		
+		@Override
+		public void run() {
+			WatchKey key;
+		    try {
+		        // wait for a key to be available
+		        key = watcher.take();
+		    } catch (InterruptedException ex) {
+		        return;
+		    }
+		    List<WatchEvent<?>> events = key.pollEvents();
+		    if (events.stream().anyMatch(event -> event.kind() == StandardWatchEventKinds.OVERFLOW)) {
+		    	propFiles.clear();
+		    	loadPrefs();
+		    } else {
+		    	for (WatchEvent<?> watchEvent : events) {
+		    		@SuppressWarnings("unchecked")
+		    		WatchEvent<Path> ev = (WatchEvent<Path>) watchEvent;
+					if (watchEvent.kind() == StandardWatchEventKinds.ENTRY_DELETE) {
+						String name = ev.context().getFileName().toString();
+						if (name.endsWith(EXT)) {
+							name = name.substring(0, name.length() - EXT.length());
+						}
+						propFiles.remove(name);
+					} else {
+						File file = ev.context().toFile();
+						if (!ev.context().isAbsolute()) {
+							file = new File(preferenceFolder, file.getName());
+						}
+						loadPrefsFromFile(file);
+					}
+				}
+		    }
+		    if (events.size() > 0) {
+		    	firePrefsChanged();
+		    }
+		 
+		    // IMPORTANT: The key must be reset after processed
+		    boolean valid = key.reset();
+		    if (!valid) {
+		        return;
+		    }
+			
+		}
+	};
+
+	private ScheduledFuture<?> scheduledFuture;
+
+	private IPreferenceUpdateCallback updateCallback;
 	
 	public FolderPreferenceProvider(String prefsFolderPath) {
 		this(new File(prefsFolderPath));
@@ -30,6 +100,12 @@ public class FolderPreferenceProvider implements IPreferenceProvider {
 		super();
 		this.preferenceFolder = preferenceFolder;
 		loadPrefs();
+		try {
+			watcher = FileSystems.getDefault().newWatchService();
+			preferenceFolder.toPath().register(watcher, StandardWatchEventKinds.ENTRY_CREATE, StandardWatchEventKinds.ENTRY_DELETE, StandardWatchEventKinds.ENTRY_MODIFY);
+		} catch (IOException e) {
+			PrefEditorPlugin.log(e);
+		}
 	}
 
 	public File getPreferenceFolder() {
@@ -42,19 +118,23 @@ public class FolderPreferenceProvider implements IPreferenceProvider {
 		}
 		File[] prefFiles = preferenceFolder.listFiles((parent, name) -> name.endsWith(EXT));
 		for (File file : prefFiles) {
-			String name = file.getName();
-			if (name.endsWith(EXT)) {
-				name = name.substring(0, name.length() - EXT.length());
-			}
-			try (InputStream stream = new BufferedInputStream(new FileInputStream(file))) {
-				Properties properties = new Properties();
-				properties.load(stream);
-				propFiles.put(name, properties);
-			} catch (FileNotFoundException e) {
-				PrefEditorPlugin.log(e);
-			} catch (IOException e) {
-				PrefEditorPlugin.log(e);
-			}
+			loadPrefsFromFile(file);
+		}
+	}
+
+	protected void loadPrefsFromFile(File file) {
+		String name = file.getName();
+		if (name.endsWith(EXT)) {
+			name = name.substring(0, name.length() - EXT.length());
+		}
+		try (InputStream stream = new BufferedInputStream(new FileInputStream(file))) {
+			Properties properties = new Properties();
+			properties.load(stream);
+			propFiles.put(name, properties);
+		} catch (FileNotFoundException e) {
+			PrefEditorPlugin.log(e);
+		} catch (IOException e) {
+			PrefEditorPlugin.log(e);
 		}
 	}
 	
@@ -139,4 +219,36 @@ public class FolderPreferenceProvider implements IPreferenceProvider {
 			PrefEditorPlugin.log("File " + category + EXT + " can't be deleted");
 		}
 	}
+
+	public void setUpdateCallback(IPreferenceUpdateCallback updateCallback) {
+		this.updateCallback = updateCallback;
+	}
+
+	public IPreferenceUpdateCallback getUpdateCallback() {
+		return updateCallback;
+	}
+
+	public void setTracking(boolean tracking) {
+		this.tracking = tracking;
+		if (tracking) {
+			scheduledFuture = scheduledExecutorService.scheduleAtFixedRate(watchRunnable , 0L, 2L, TimeUnit.SECONDS);
+		} else {
+			scheduledFuture.cancel(true);
+		}
+	}
+
+	public boolean isTracking() {
+		return tracking;
+	}
+
+	protected void firePrefsChanged() {
+		updateCallback.preferencesUpdated("");
+	}
+
+	@Override
+	public void dispose() {
+		scheduledFuture.cancel(false);
+		scheduledExecutorService.shutdown();
+	}
+
 }