+++ plugins/com.onpositive.prefeditor/src/com/onpositive/prefeditor/model/IPreferenceProvider.java
@@ -15,4 +15,12 @@ public interface IPreferenceProvider {
 
 	public abstract String[] getNodeNames();
 
+	public abstract void setUpdateCallback(IPreferenceUpdateCallback updateCallback);
+
+	public abstract IPreferenceUpdateCallback getUpdateCallback();
+
+	public abstract void setTracking(boolean tracking);
+
+	public abstract boolean isTracking();
+
 }